+++ js/changes.md
@@ -9,6 +9,7 @@
 6. Support different marker styles in 3D drawings
 7. Implement interactive movement of TArrow class
 8. New and simpler TArrow drawing without use of svg markers
+9. Use TGeoManager::fVisLevel in geometry painter
 
 
 ## Changes in 5.7.0